+++ App.jsx
@@ -1,23 +1,9 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
 
-import AppBar from './src/components/AppBar';
-import MemoList from './src/components/MemoList';
-import CircleButton from './src/components/CircleButton';
+import MemoListScreen from './src/screens/MemoListScreen';
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <AppBar />
-      <MemoList />
-      <CircleButton>+</CircleButton>
-    </View>
+    <MemoListScreen />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f0f4f8',
-  },
-});
