+++ src/screens/MemoListScreen.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { View, StyleSheet } from 'react-native';
+
+import AppBar from '../components/AppBar';
+import MemoList from '../components/MemoList';
+import CircleButton from '../components/CircleButton';
+
+export default function MemoListScreen() {
+  return (
+    <View style={styles.container}>
+      <AppBar />
+      <MemoList />
+      <CircleButton>+</CircleButton>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#f0f4f8',
+  },
+});
